requirements.txt
# INS Dashboard - Python Dependencies
# Generated: October 23, 2025

# ============================================================================
# CORE DEPENDENCIES (Production)
# ============================================================================

# Database & API
supabase>=2.0.0,<3.0.0          # Supabase client for database operations
requests>=2.31.0,<3.0.0         # HTTP requests for API calls
python-dotenv>=1.0.0,<2.0.0     # Environment variable management

# Data Processing
pandas>=2.0.0,<3.0.0            # Data manipulation and analysis
numpy>=1.24.0,<2.0.0            # Numerical computing

# FIT File Processing
fitparse>=1.2.0,<2.0.0          # Garmin FIT file parsing

# Dashboard & Visualization
shiny>=0.6.0,<1.0.0             # Interactive web applications
plotly>=5.15.0,<6.0.0           # Interactive plotting
matplotlib>=3.7.0,<4.0.0        # Static plotting (fallback)

# Date/Time Handling
python-dateutil>=2.8.0,<3.0.0   # Extended date/time utilities

# ============================================================================
# DEVELOPMENT DEPENDENCIES (Optional)
# ============================================================================

# Testing
pytest>=7.4.0,<8.0.0            # Testing framework
pytest-cov>=4.1.0,<5.0.0        # Test coverage reporting
pytest-mock>=3.11.0,<4.0.0      # Mocking for tests

# Code Quality
black>=23.7.0,<24.0.0           # Code formatting
flake8>=6.0.0,<7.0.0            # Linting
mypy>=1.5.0,<2.0.0              # Type checking
isort>=5.12.0,<6.0.0            # Import sorting

# Documentation
sphinx>=7.1.0,<8.0.0            # Documentation generation
sphinx-rtd-theme>=1.3.0,<2.0.0  # Read the Docs theme

# ============================================================================
# SYSTEM DEPENDENCIES (Usually pre-installed)
# ============================================================================

# Standard library dependencies (included with Python)
# - os, sys, json, datetime, time, argparse
# - typing, pathlib, ast, re, io, csv
# - unittest.mock, traceback

# ============================================================================
# INSTALLATION NOTES
# ============================================================================

# Install production dependencies only:
# pip install -r requirements.txt

# Install with development dependencies:
# pip install -r requirements.txt
# pip install pytest black flake8 mypy

# For specific Python versions:
# Python 3.11+ recommended
# Minimum Python 3.9 supported

# ============================================================================
# VERSION COMPATIBILITY
# ============================================================================

# Tested with:
# - Python 3.11.5
# - Python 3.12.0
# - macOS 13+ (Apple Silicon & Intel)
# - Ubuntu 20.04+ LTS
# - Windows 10+ (via WSL recommended)

# ============================================================================
# OPTIONAL SYSTEM PACKAGES
# ============================================================================

# For better performance (optional):
# - psycopg2-binary>=2.9.0  # PostgreSQL adapter (if using direct DB access)
# - uvloop>=0.17.0          # Fast event loop (Unix only)

# For development environment:
# - jupyter>=1.0.0          # Notebook environment for data exploration
# - ipython>=8.0.0          # Enhanced Python shell
